requirements.txt
MySQL-python==1.2.5
PyYAML==3.10
argparse==1.2.1
cffi==0.8
greenlet==0.4.0
mock==1.0.1
nltk==2.0.4
nose==1.3.3
-e git+https://bitbucket.org/pypy/numpy.git@198614f1bad70ded961ad16c77ddf972cf9dda01#egg=numpy-master
oauthlib==0.6.1
python-dateutil==2.2
readline==6.2.4.1
redis==2.10.1
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
wsgiref==0.1.2
